web_project/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 环境变量
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 日期时间
python-dateutil==2.8.2
